requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
langchain==0.1.0
langgraph==0.0.26
openai==1.12.0
chromadb==0.4.22
pydantic==2.6.1
python-socketio==5.10.0
llama-cpp-python==0.2.57
sentence-transformers==2.3.1
transformers==4.36.1
torch==2.0.1
serpapi==2.0.0
google-generative-ai==0.4.0
Flask==3.0.0
beautifulsoup4==4.12.0
requests==2.31.0
google-api-core==2.17.0
protobuf==4.29.1
